+++ Assets/A_Scripts/Entity.cs
@@ -8,14 +8,14 @@ public abstract class Entity : MonoBehaviour
     [SerializeField] public enum Behavior {Neutral, Offense, Defense};
     [SerializeField] public Behavior behavior;
 
-    [SerializeField] protected int hp;
-    [SerializeField] protected int maxHP;
-    [SerializeField] protected int def;
-    [SerializeField] protected int speed;
-    [SerializeField] protected int attackSpeed;
-    [SerializeField] protected int attack;
-    [SerializeField] protected int mana;
-    [SerializeField] protected int maxMana;
+    [SerializeField] protected float hp;
+    [SerializeField] protected float maxHP;
+    [SerializeField] protected float def;
+    [SerializeField] protected float speed;
+    [SerializeField] protected float attackSpeed;
+    [SerializeField] protected float attack;
+    [SerializeField] protected float mana;
+    [SerializeField] protected float maxMana;
     [SerializeField] protected int cost;
     [SerializeField] protected float range;
 
@@ -31,31 +31,31 @@ public abstract class Entity : MonoBehaviour
    
 
     // PV
-    public abstract int GetHP();
-    public abstract void SetHP(int newHP);
+    public abstract float GetHP();
+    public abstract void SetHP(float newHP);
 
-    public abstract int GetMaxHP();
-    public abstract void SetMaxHP(int newMaxHP);  
+    public abstract float GetMaxHP();
+    public abstract void SetMaxHP(float newMaxHP);  
 
     // Défense
-    public abstract int GetDef();
-    public abstract void SetDef(int newDef);
+    public abstract float GetDef();
+    public abstract void SetDef(float newDef);
 
     // Vitesse
-    public abstract int GetSpeed(); // La vitesse de déplacement
-    public abstract void SetSpeed(int newSpeed);
+    public abstract float GetSpeed(); // La vitesse de déplacement
+    public abstract void SetSpeed(float newSpeed);
 
     // Attaque
-    public abstract int GetAttackSpeed();
-    public abstract void SetAttackSpeed(int newAttackSpeed);
-    public abstract int GetAttack();
-    public abstract void SetAttack(int newAttack);
+    public abstract float GetAttackSpeed();
+    public abstract void SetAttackSpeed(float newAttackSpeed);
+    public abstract float GetAttack();
+    public abstract void SetAttack(float newAttack);
 
     // Mana
-    public abstract int GetMana();
-    public abstract void SetMana(int newMana);
-    public abstract int GetMaxMana();
-    public abstract void SetMaxMana(int newMaxMana);
+    public abstract float GetMana();
+    public abstract void SetMana(float newMana);
+    public abstract float GetMaxMana();
+    public abstract void SetMaxMana(float newMaxMana);
 
     // Coût
     public abstract int GetCost(); // Le coût à payer pour placer l'entity sur l'arène
@@ -63,7 +63,7 @@ public abstract class Entity : MonoBehaviour
 
     // Portée
     public abstract float GetRange(); 
-    public abstract void SetRange(int newRange);
+    public abstract void SetRange(float newRange);
 
     // Pour les capacités spéciales : Les différentes capacités sont des components qui héritent de la classe "Capacité".
     // => Pour donner une capacité spéciale à une unité il faut lui donner le component (script) qui correspond à la capacité spéciale choisie
